@@ -1,6 +1,4 @@
 const { defineConfig } = require('@vue/cli-service')
 module.exports = defineConfig({
-  transpileDependencies: [
-    'vuetify'
-  ]
+  transpileDependencies: true
 })
